requirements.txt
Flask==3.1.1
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_core==0.3.74
langchain_mistralai==0.2.11
langchain_ollama==0.3.6
langgraph==0.6.4
mysql_connector_repackaged==0.3.1
Requests==2.32.4
streamlit==1.47.1
typing_extensions==4.14.1
